requirements.txt
requests==2.31.0
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
web3==6.15.1
solana==0.30.2
schedule==1.2.0
